+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.5.1 - 2021-10-03
+
+* Fixed: Set `max-width` for DB error logo image.
+
 ## 1.5.0 - 2021-10-02
 
 * New: Customized Database Error page.